[CLOUD-2596] - Additional line for Statefulsets looking up image streams
added "alpha.image.policy.openshift.io/resolve-names: '*'" to the stateful set templates
It isn't required in the openshift instance I was testing on.  But according to the documentation, this line could prevent the need for local lookup

amq-broker-71-statefulset.yaml was an early version of templates/amq-broker-71-persistence.yaml, so I have removed it.

Signed-off-by: Vanessa Busch <vbusch@redhat.com>