ENTMQBR-1648 Don't hardcode openshift.DNS_PING service name
Now the service name takes the form of
${APPLICATION_NAME}-${PING_SVC_NAME}

If both APPLICATION_NAME and PING_SVC_NAME are absent the
DNS_PING service name will be default to 'ping' (that's the
original value). If only APPLICATION_NAME is absent, the service
name is the value of PING_SVC_NAME (without prefixed with
the dash).